Data frame or conversion factors for length or areal units

Data frame of conversion factors for length or areal units.

Usage

conversionFactors

Format

An object of class 'data.frame'.

See also

Examples

data(conversionFactors)
conversionFactors
#>                from              to       factor
#> 1             acres           acres 1.000000e+00
#> 2             acres           feet2 4.356000e+04
#> 3             acres        hectares 4.046860e-01
#> 4             acres     kilometers2 4.046860e-03
#> 5             acres         meters2 4.046860e+03
#> 6             acres          miles2 1.562500e-03
#> 7             acres nautical miles2 1.179875e-03
#> 8             acres          yards2 4.840000e+03
#> 9              feet            feet 1.000000e+00
#> 10             feet      kilometers 3.048000e-04
#> 11             feet          meters 3.048000e-01
#> 12             feet           miles 1.893940e-04
#> 13             feet  nautical miles 1.645790e-04
#> 14             feet           yards 3.333333e-01
#> 15            feet2           acres 2.300000e-05
#> 16            feet2           feet2 1.000000e+00
#> 17            feet2        hectares 9.290000e-06
#> 18            feet2     kilometers2 9.290000e-08
#> 19            feet2         meters2 9.290304e-02
#> 20            feet2          miles2 3.590000e-08
#> 21            feet2 nautical miles2 2.710000e-08
#> 22            feet2          yards2 1.111111e-01
#> 23         hectares           acres 2.471050e+00
#> 24         hectares           feet2 1.076390e+05
#> 25         hectares        hectares 1.000000e+00
#> 26         hectares     kilometers2 1.000000e-02
#> 27         hectares         meters2 1.000000e+04
#> 28         hectares          miles2 3.861020e-03
#> 29         hectares nautical miles2 3.429904e+02
#> 30         hectares          yards2 1.195990e+04
#> 31       kilometers            feet 3.280840e+03
#> 32       kilometers      kilometers 1.000000e+00
#> 33       kilometers          meters 1.000000e-03
#> 34       kilometers           miles 6.213710e-01
#> 35       kilometers  nautical miles 5.399570e-01
#> 36       kilometers           yards 1.093610e+03
#> 37      kilometers2           acres 2.471050e+02
#> 38      kilometers2           feet2 1.080000e+07
#> 39      kilometers2        hectares 1.000000e+02
#> 40      kilometers2     kilometers2 1.000000e+00
#> 41      kilometers2         meters2 1.000000e+06
#> 42      kilometers2          miles2 3.861020e-01
#> 43      kilometers2 nautical miles2 2.915530e-01
#> 44      kilometers2          yards2 1.200000e+06
#> 45           meters            feet 3.280840e+00
#> 46           meters      kilometers 1.000000e-03
#> 47           meters          meters 1.000000e+00
#> 48           meters           miles 6.213710e-04
#> 49           meters  nautical miles 5.399570e-04
#> 50           meters           yards 1.093613e+00
#> 51          meters2           acres 2.471050e-04
#> 52          meters2           feet2 1.076390e+01
#> 53          meters2        hectares 1.000000e-04
#> 54          meters2     kilometers2 1.000000e-06
#> 55          meters2         meters2 1.000000e+00
#> 56          meters2          miles2 3.860000e-07
#> 57          meters2 nautical miles2 2.920000e-07
#> 58          meters2          yards2 1.195990e+00
#> 59            miles            feet 5.280000e+03
#> 60            miles      kilometers 1.609344e+00
#> 61            miles          meters 1.609340e+03
#> 62            miles           miles 1.000000e+00
#> 63            miles  nautical miles 8.689760e-01
#> 64            miles           yards 1.760000e+03
#> 65           miles2           acres 6.400000e+02
#> 66           miles2           feet2 2.790000e+07
#> 67           miles2        hectares 2.589990e+02
#> 68           miles2     kilometers2 2.589988e+00
#> 69           miles2         meters2 2.590000e+06
#> 70           miles2          miles2 1.000000e+00
#> 71           miles2 nautical miles2 1.324293e+00
#> 72           miles2          yards2 3.100000e+06
#> 73   nautical miles            feet 6.076120e+03
#> 74   nautical miles      kilometers 1.852000e+00
#> 75   nautical miles          meters 1.852000e+03
#> 76   nautical miles           miles 1.150780e+00
#> 77   nautical miles  nautical miles 1.000000e+00
#> 78   nautical miles           yards 2.025370e+03
#> 79  nautical miles2           acres 1.179875e-03
#> 80  nautical miles2           feet2 3.691918e+07
#> 81  nautical miles2        hectares 3.429904e+02
#> 82  nautical miles2     kilometers2 3.429900e+00
#> 83  nautical miles2         meters2 3.400000e+06
#> 84  nautical miles2          miles2 1.324293e+00
#> 85  nautical miles2 nautical miles2 1.000000e+00
#> 86  nautical miles2          yards2 2.440000e-07
#> 87            yards            feet 3.000000e+00
#> 88            yards      kilometers 9.144000e-04
#> 89            yards          meters 9.144000e-01
#> 90            yards           miles 5.681820e-04
#> 91            yards  nautical miles 4.937370e-04
#> 92            yards           yards 1.000000e+00
#> 93           yards2           acres 2.066120e-04
#> 94           yards2           feet2 9.000000e+00
#> 95           yards2        hectares 8.360000e-05
#> 96           yards2     kilometers2 8.360000e-07
#> 97           yards2         meters2 8.361274e-01
#> 98           yards2          miles2 3.230000e-07
#> 99           yards2 nautical miles2 2.440000e-07
#> 100          yards2          yards2 1.000000e+00