Calculate pairwise distances between two matrices or data frames.

This function takes two data frames or matrices and returns a matrix of pairwise Euclidean distances between the two.

Usage

pairDist(x1, x2 = NULL, na.rm = FALSE)

Arguments

x1

Data frame or matrix one or more columns wide.

x2

Data frame or matrix one or more columns wide. If NULL, then pairwise distances between all points in x1 are calculated.

na.rm

Logical, if TRUE then any rows in x1 or x2 with at least one NA are removed first.

Value

Matrix with nrow(x1) rows and nrow(x2) columns. Values are the distance between each row of x1 and row of x2.

See also

Examples

x1 <- data.frame(x=sample(1:30, 30), y=sort(round(100 * rnorm(30))), z=sample(1:30, 30))
x2 <- data.frame(x=1:20, y=round(100 * rnorm(20)), z=sample(1:20, 20))
pairDist(x1, x2)
#>            [,1]      [,2]       [,3]      [,4]       [,5]      [,6]      [,7]
#>  [1,]  57.21014  55.39856  83.701852 281.51554  27.221315 114.20595 321.43117
#>  [2,]  51.89412  50.30904  79.271685 278.14565  20.712315 110.38569 318.15719
#>  [3,]  50.13980  50.67544  76.491830 275.57576  17.146428 107.78683 315.20628
#>  [4,]  46.87217  46.78675  74.390860 274.15507  12.688578 106.08016 314.01592
#>  [5,]  41.35215  41.95235  69.072426 269.15795   6.480741 101.04454 309.00324
#>  [6,]  27.67671  21.77154  53.075418 249.21878  27.531800  82.57118 289.56001
#>  [7,]  14.73092  14.79865  40.890097 240.08540  25.942244  72.09022 280.03571
#>  [8,]  18.05547  14.00000  34.336569 229.33382  40.024992  62.46599 269.37520
#>  [9,]  24.81935  18.38478  33.481338 222.65220  49.839743  57.67148 262.87830
#> [10,]  36.04164  33.42155  33.256578 212.68051  61.603571  50.68530 252.44009
#> [11,]  35.58089  33.61547  30.870698 210.52791  62.369865  48.06246 250.25787
#> [12,]  22.15852  25.31798   8.831761 207.19315  57.297469  39.10243 247.00202
#> [13,]  35.88872  35.69314  23.853721 203.09111  66.241981  39.26831 242.76944
#> [14,]  39.96248  37.37646  26.944387 199.02010  71.617037  37.16181 239.07739
#> [15,]  34.13210  37.34970   6.324555 194.28072  70.178344  26.77686 234.08545
#> [16,]  39.93745  38.32754  20.149442 191.17793  76.019734  27.22132 231.60743
#> [17,]  51.77837  54.87258  27.568098 183.68723  84.005952  24.18677 222.71955
#> [18,]  53.48832  56.54202  26.419690 177.97472  87.595662  15.58846 217.26021
#> [19,]  77.54998  76.45914  52.583267 156.18579 112.703150  25.72936 196.38992
#> [20,]  88.24398  90.42677  60.149813 142.04577 123.332883  29.27456 181.23465
#> [21,]  90.16097  91.14275  62.513998 138.01811 126.495059  30.41381 178.14601
#> [22,] 105.66456 105.03809  79.334734 127.67537 141.226768  48.17676 167.82729
#> [23,] 133.99254 135.10736 106.174385  98.24459 169.215839  74.21590 137.07297
#> [24,] 138.11589 139.16178 110.240646  90.05554 174.287119  78.06408 130.11149
#> [25,] 201.11191 202.64254 173.028899  29.96665 237.002110 141.18782  67.25325
#> [26,] 209.70694 210.23083 181.972525  21.42429 245.888186 149.64959  61.00820
#> [27,] 244.50358 245.51171 216.506351  28.74022 280.060708 184.32037  32.78719
#> [28,] 246.41023 247.20437 218.515446  19.79899 282.485398 186.26862  24.67793
#> [29,] 279.37967 280.03750 251.503479  54.85435 315.215799 219.19170  25.09980
#> [30,] 299.17386 300.41804 271.071946  74.65253 334.725260 239.00000  36.27671
#>            [,8]       [,9]      [,10]     [,11]     [,12]      [,13]     [,14]
#>  [1,] 122.38872  13.341664  66.678332 180.98619  93.56816  75.531450  97.25739
#>  [2,] 118.71394   9.165151  62.721607 177.64571  90.09439  72.062473  94.02659
#>  [3,] 115.43830  16.822604  59.573484 174.18668  87.48714  69.641941  92.01087
#>  [4,] 114.12712  13.038405  58.034473 173.18776  86.07555  68.183576  90.53729
#>  [5,] 109.05962  18.627936  53.150729 168.18145  81.37567  63.671030  86.08136
#>  [6,]  91.68424  38.948684  38.196859 150.14993  62.78535  45.321077  65.86350
#>  [7,]  80.40522  46.043458  24.535688 139.44892  52.08647  34.249088  56.54202
#>  [8,]  71.06335  57.645468  18.681542 129.35610  41.78516  23.958297  45.05552
#>  [9,]  66.73829  65.795137  22.022716 123.90722  37.49667  21.633308  39.16631
#> [10,]  58.38664  78.294317  25.495098 113.58697  30.87070  19.874607  31.63858
#> [11,]  55.65968  79.780950  24.062419 111.12605  28.17801  17.720045  29.22328
#> [12,]  47.16991  79.233831   9.273618 106.27323  19.82423   7.141428  25.78759
#> [13,]  46.62617  85.516080  22.022716 102.92230  19.33908  12.806248  21.30728
#> [14,]  45.61798  89.632583  28.142495 100.52860  19.20937  17.000000  17.57840
#> [15,]  34.77068  92.693042  23.790755  93.66963  13.89244  18.357560  20.32240
#> [16,]  37.37646  95.781000  31.272992  93.32738  15.90597  21.610183  15.13275
#> [17,]  26.92582 105.867842  37.067506  81.89017  17.34935  29.068884  20.22375
#> [18,]  18.62794 109.936345  39.774364  76.10519  15.84298  31.448370  18.41195
#> [19,]  26.87006 132.124941  65.138314  60.90156  37.33631  53.497664  31.08054
#> [20,]  20.02498 145.581592  75.272837  40.12481  48.09366  65.931783  45.88028
#> [21,]  24.67793 148.121572  78.625696  40.02499  50.84290  68.738635  47.46578
#> [22,]  43.87482 161.077621  93.187982  38.60052  64.78426  81.884064  59.33801
#> [23,]  66.24198 190.633156 120.851148  16.58312  92.80086 110.589330  88.92694
#> [24,]  70.63993 196.043363 126.241831  17.34935  98.25477 116.275535  94.56215
#> [25,] 133.06389 259.281700 189.137516  74.35725 161.44968 179.504875 158.03164
#> [26,] 142.30601 267.159129 197.651208  84.89994 169.36056 187.277868 165.10300
#> [27,] 176.27819 301.613992 231.734762 117.66478 203.63202 221.517494 199.61463
#> [28,] 178.59451 304.041116 234.266942 120.21231 206.08736 224.058028 202.00990
#> [29,] 211.45449 336.502600 266.848271 153.05555 238.58122 256.454674 234.32670
#> [30,] 230.78345 356.595008 286.520505 171.77311 258.55947 276.501356 254.71749
#>            [,15]     [,16]     [,17]     [,18]     [,19]     [,20]
#>  [1,] 140.641388 291.04295 122.21293 239.62262 194.58931 120.02083
#>  [2,] 137.455447 288.01736 119.17214 236.59459 191.60637 117.21348
#>  [3,] 134.033578 285.25953 116.15937 233.05364 188.02659 114.45523
#>  [4,] 133.154046 284.17600 115.23020 232.31229 187.34727 113.60018
#>  [5,] 128.331602 279.35819 110.62097 227.42471 182.53219 109.19707
#>  [6,] 110.313191 259.31255  91.78780 208.91625 164.20719  89.63258
#>  [7,]  99.428366 250.14796  81.40025 198.56737 153.65220  79.81854
#>  [8,]  89.112289 239.00837  70.46276 187.97340 143.04545  68.23489
#>  [9,]  83.958323 232.21757  65.00000 182.01648 137.30987  62.24147
#> [10,]  73.579889 221.49944  54.71746 170.84789 126.08331  51.46844
#> [11,]  71.028163 219.40602  52.20153 168.55266 123.72550  49.09175
#> [12,]  66.437941 217.33384  48.92852 165.49622 120.65239  48.05206
#> [13,]  62.617889 212.20038  43.92038 160.85397 115.88788  41.30375
#> [14,]  60.811183 208.24025  41.73727 158.04113 113.39753  38.37968
#> [15,]  54.827001 204.84628  38.80722 153.06535 108.57716  39.12800
#> [16,]  55.045436 201.60853  37.65634 151.76627 107.65222  36.12478
#> [17,]  42.071368 192.87820  26.19160 140.08926  95.10520  25.98076
#> [18,]  36.221541 187.68591  20.68816 135.03703  90.09439  21.84033
#> [19,]  27.221315 165.39347  17.57840 116.42165  72.80110  14.35270
#> [20,]   3.162278 151.75638  20.34699  99.10096  54.23099  24.59675
#> [21,]  15.297059 148.60013  25.41653  98.16822  54.78138  28.61818
#> [22,]  28.740216 136.58697  37.70942  88.88757  47.44470  37.69615
#> [23,]  46.162756 106.53169  63.39558  55.24491  13.03840  65.31462
#> [24,]  52.583267 100.68267  70.06426  51.40039  18.05547  72.59477
#> [25,] 114.546934  41.34005 132.87964  20.73644  61.85467 135.48432
#> [26,] 123.385575  30.03331 140.64139  33.27161  71.61704 142.50965
#> [27,] 156.655673  14.31782 174.34735  58.90671 102.79105 176.22713
#> [28,] 159.558767   9.00000 177.20892  62.92853 106.44247 179.20379
#> [29,] 191.984374  41.42463 209.43257  94.70480 138.48827 211.17055
#> [30,] 211.343323  62.36986 229.32510 112.43220 157.21641 231.36767
pairDist(x1)
#>             [,1]       [,2]      [,3]       [,4]       [,5]      [,6]      [,7]
#>  [1,]   0.000000   8.602325  13.74773  15.297059  21.656408  36.06938  43.01163
#>  [2,]   8.602325   0.000000  12.44990   8.602325  14.525839  31.25700  38.41875
#>  [3,]  13.747727  12.449900   0.00000   9.000000  13.564660  35.18522  36.81032
#>  [4,]  15.297059   8.602325   9.00000   0.000000   6.708204  30.38092  34.14674
#>  [5,]  21.656408  14.525839  13.56466   6.708204   0.000000  27.60435  29.58040
#>  [6,]  36.069378  31.256999  35.18522  30.380915  27.604347   0.00000  16.76305
#>  [7,]  43.011626  38.418745  36.81032  34.146742  29.580399  16.76305   0.00000
#>  [8,]  52.278102  49.183331  48.12484  46.486557  42.941821  23.02173  14.79865
#>  [9,]  59.974995  57.210139  57.39338  55.470713  52.249402  28.80972  24.75884
#> [10,]  70.753092  69.382995  67.29785  67.364679  64.536811  44.59821  37.17526
#> [11,]  72.505172  70.943640  68.57113  68.607580  65.559134  46.04346  37.74917
#> [12,]  75.670338  71.624018  68.79680  67.037303  62.040309  45.70558  33.25658
#> [13,]  79.126481  76.948034  74.01351  73.790243  70.099929  51.51699  41.34005
#> [14,]  83.390647  81.006173  79.65551  78.587531  74.973329  53.16954  45.84757
#> [15,]  89.944427  85.440037  82.77077  80.560536  75.173133  58.83026  47.07441
#> [16,]  92.249661  88.056800  87.59566  84.723078  79.931220  58.24946  50.93133
#> [17,] 100.468901  97.836598  93.26843  93.348808  88.921314  73.62744  60.86050
#> [18,] 105.394497 102.127371  98.12747  97.457683  92.633687  76.63550  64.23395
#> [19,] 126.495059 123.729544 122.18838 120.842873 116.644760  94.78396  87.06894
#> [20,] 140.975175 137.717101 134.09325 133.244137 128.347185 110.85576  99.62931
#> [21,] 144.159634 140.370937 138.11227 136.271787 131.221187 111.71840 102.19589
#> [22,] 155.467038 152.715422 150.93376 149.679658 145.358178 123.72146 115.78428
#> [23,] 185.156690 182.430809 179.17031 178.552513 173.959765 154.77080 144.79296
#> [24,] 191.710720 188.194049 185.62866 184.111379 179.103322 159.54310 150.05666
#> [25,] 254.952937 251.497515 248.40290 247.145706 242.041319 223.20842 213.24868
#> [26,] 262.259414 259.102296 256.72748 255.385982 250.561370 230.05434 221.28262
#> [27,] 296.143546 293.426311 290.23439 289.566918 284.878922 265.25836 255.70882
#> [28,] 299.137092 296.008446 293.34962 292.145512 287.273389 267.16661 258.08719
#> [29,] 331.061928 328.295599 325.49194 324.616697 319.932805 299.64479 290.63723
#> [30,] 351.291901 348.488163 345.11882 344.456093 339.663068 320.38258 310.62518
#>            [,8]      [,9]     [,10]      [,11]     [,12]      [,13]     [,14]
#>  [1,]  52.27810  59.97499  70.75309  72.505172  75.67034  79.126481  83.39065
#>  [2,]  49.18333  57.21014  69.38300  70.943640  71.62402  76.948034  81.00617
#>  [3,]  48.12484  57.39338  67.29785  68.571131  68.79680  74.013512  79.65551
#>  [4,]  46.48656  55.47071  67.36468  68.607580  67.03730  73.790243  78.58753
#>  [5,]  42.94182  52.24940  64.53681  65.559134  62.04031  70.099929  74.97333
#>  [6,]  23.02173  28.80972  44.59821  46.043458  45.70558  51.516987  53.16954
#>  [7,]  14.79865  24.75884  37.17526  37.749172  33.25658  41.340053  45.84757
#>  [8,]   0.00000  10.67708  23.00000  23.874673  25.86503  28.740216  32.14032
#>  [9,]  10.67708   0.00000  17.74824  19.026298  25.82634  24.657656  24.59675
#> [10,]  23.00000  17.74824   0.00000   3.000000  26.38181  12.449900  15.55635
#> [11,]  23.87467  19.02630   3.00000   0.000000  24.26932   9.486833  13.96424
#> [12,]  25.86503  25.82634  26.38181  24.269322   0.00000  19.052559  23.79075
#> [13,]  28.74022  24.65766  12.44990   9.486833  19.05256   0.000000  11.18034
#> [14,]  32.14032  24.59675  15.55635  13.964240  23.79075  11.180340   0.00000
#> [15,]  40.23680  38.48376  37.22902  34.713110  14.89966  26.851443  28.60070
#> [16,]  40.80441  34.56877  34.46738  32.695565  23.40940  27.037012  21.21320
#> [17,]  51.17617  48.63127  37.06751  34.132096  30.23243  25.079872  29.69848
#> [18,]  55.03635  52.00961  42.68489  39.761791  31.59114  30.347982  32.64966
#> [19,]  74.60563  67.03730  57.38467  55.497748  56.78908  49.132474  43.20880
#> [20,]  89.64932  84.82335  74.53858  71.923571  66.54322  63.332456  61.98387
#> [21,]  92.05976  86.14523  78.24321  75.848533  69.44062  67.683085  63.89053
#> [22,] 103.56158  96.06768  85.87200  83.970233  84.53402  77.440300  72.13876
#> [23,] 133.54400 127.33421 115.77996 113.551750 112.16506 106.122571 103.11644
#> [24,] 139.56361 133.28166 123.88301 121.597697 117.14521 113.789279 109.98636
#> [25,] 202.98768 196.90099 186.79668 184.493902 180.11385 176.663522 173.34647
#> [26,] 210.09760 203.12804 192.94041 190.874304 188.62131 183.709009 179.34882
#> [27,] 244.42586 237.79403 226.27638 224.182961 222.91927 217.057596 213.54391
#> [28,] 247.03441 240.22073 229.74116 227.644460 225.25763 220.404174 216.34463
#> [29,] 279.18632 272.21499 261.03257 259.019304 257.95736 252.073402 248.06451
#> [30,] 299.57804 293.07508 281.59190 279.465561 277.67967 272.229682 268.86800
#>           [,15]     [,16]      [,17]      [,18]     [,19]     [,20]     [,21]
#>  [1,]  89.94443  92.24966 100.468901 105.394497 126.49506 140.97518 144.15963
#>  [2,]  85.44004  88.05680  97.836598 102.127371 123.72954 137.71710 140.37094
#>  [3,]  82.77077  87.59566  93.268430  98.127468 122.18838 134.09325 138.11227
#>  [4,]  80.56054  84.72308  93.348808  97.457683 120.84287 133.24414 136.27179
#>  [5,]  75.17313  79.93122  88.921314  92.633687 116.64476 128.34719 131.22119
#>  [6,]  58.83026  58.24946  73.627441  76.635501  94.78396 110.85576 111.71840
#>  [7,]  47.07441  50.93133  60.860496  64.233947  87.06894  99.62931 102.19589
#>  [8,]  40.23680  40.80441  51.176166  55.036352  74.60563  89.64932  92.05976
#>  [9,]  38.48376  34.56877  48.631266  52.009614  67.03730  84.82335  86.14523
#> [10,]  37.22902  34.46738  37.067506  42.684892  57.38467  74.53858  78.24321
#> [11,]  34.71311  32.69557  34.132096  39.761791  55.49775  71.92357  75.84853
#> [12,]  14.89966  23.40940  30.232433  31.591138  56.78908  66.54322  69.44062
#> [13,]  26.85144  27.03701  25.079872  30.347982  49.13247  63.33246  67.68309
#> [14,]  28.60070  21.21320  29.698485  32.649655  43.20880  61.98387  63.89053
#> [15,]   0.00000  18.27567  25.768197  22.583180  48.25971  54.71746  56.60389
#> [16,]  18.27567   0.00000  33.136083  29.899833  40.01250  55.92853  53.94442
#> [17,]  25.76820  33.13608   0.000000   9.486833  39.10243  42.02380  50.31898
#> [18,]  22.58318  29.89983   9.486833   0.000000  35.05710  36.02777  42.82523
#> [19,]  48.25971  40.01250  39.102430  35.057096   0.00000  29.94996  27.22132
#> [20,]  54.71746  55.92853  42.023803  36.027767  29.94996   0.00000  16.55295
#> [21,]  56.60389  53.94442  50.318983  42.825226  27.22132  16.55295   0.00000
#> [22,]  74.25631  67.63135  63.134776  58.395205  29.03446  31.46427  25.65151
#> [23,] 100.74225  98.51396  86.585218  81.993902  61.88699  46.78675  47.12749
#> [24,] 104.31203 101.64153  94.641429  88.391176  68.16157  52.94337  48.05206
#> [25,] 167.05388 165.35114 156.188988 150.416090 131.27833 114.44213 111.54820
#> [26,] 176.10792 172.23821 165.051507 159.574434 136.32681 123.83053 119.60769
#> [27,] 210.79611 208.04086 197.370211 192.699248 171.08477 156.88531 155.05805
#> [28,] 212.62408 209.32511 201.109423 195.762611 173.43010 159.81552 156.31059
#> [29,] 245.73563 242.21065 232.967809 228.144691 205.23889 192.31745 189.53100
#> [30,] 265.28475 262.95627 252.063484 247.309523 226.41776 211.40956 209.61870
#>           [,22]     [,23]     [,24]     [,25]     [,26]     [,27]     [,28]
#>  [1,] 155.46704 185.15669 191.71072 254.95294 262.25941 296.14355 299.13709
#>  [2,] 152.71542 182.43081 188.19405 251.49751 259.10230 293.42631 296.00845
#>  [3,] 150.93376 179.17031 185.62866 248.40290 256.72748 290.23439 293.34962
#>  [4,] 149.67966 178.55251 184.11138 247.14571 255.38598 289.56692 292.14551
#>  [5,] 145.35818 173.95977 179.10332 242.04132 250.56137 284.87892 287.27339
#>  [6,] 123.72146 154.77080 159.54310 223.20842 230.05434 265.25836 267.16661
#>  [7,] 115.78428 144.79296 150.05666 213.24868 221.28262 255.70882 258.08719
#>  [8,] 103.56158 133.54400 139.56361 202.98768 210.09760 244.42586 247.03441
#>  [9,]  96.06768 127.33421 133.28166 196.90099 203.12804 237.79403 240.22073
#> [10,]  85.87200 115.77996 123.88301 186.79668 192.94041 226.27638 229.74116
#> [11,]  83.97023 113.55175 121.59770 184.49390 190.87430 224.18296 227.64446
#> [12,]  84.53402 112.16506 117.14521 180.11385 188.62131 222.91927 225.25763
#> [13,]  77.44030 106.12257 113.78928 176.66352 183.70901 217.05760 220.40417
#> [14,]  72.13876 103.11644 109.98636 173.34647 179.34882 213.54391 216.34463
#> [15,]  74.25631 100.74225 104.31203 167.05388 176.10792 210.79611 212.62408
#> [16,]  67.63135  98.51396 101.64153 165.35114 172.23821 208.04086 209.32511
#> [17,]  63.13478  86.58522  94.64143 156.18899 165.05151 197.37021 201.10942
#> [18,]  58.39521  81.99390  88.39118 150.41609 159.57443 192.69925 195.76261
#> [19,]  29.03446  61.88699  68.16157 131.27833 136.32681 171.08477 173.43010
#> [20,]  31.46427  46.78675  52.94337 114.44213 123.83053 156.88531 159.81552
#> [21,]  25.65151  47.12749  48.05206 111.54820 119.60769 155.05805 156.31059
#> [22,]   0.00000  35.73514  42.15448 103.60019 107.54534 142.36924 144.64093
#> [23,]  35.73514   0.00000  20.14944  72.18033  79.16439 111.07655 114.90866
#> [24,]  42.15448  20.14944   0.00000  63.79655  71.95137 107.50814 108.34205
#> [25,] 103.60019  72.18033  63.79655   0.00000  22.64950  49.03060  48.10405
#> [26,] 107.54534  79.16439  71.95137  22.64950   0.00000  39.86226  37.34970
#> [27,] 142.36924 111.07655 107.50814  49.03060  39.86226   0.00000  17.72005
#> [28,] 144.64093 114.90866 108.34205  48.10405  37.34970  17.72005   0.00000
#> [29,] 176.28386 146.42746 141.79210  82.07923  70.63993  36.45545  35.05710
#> [30,] 197.71191 166.13549 161.75599  99.69453  92.40130  55.45268  55.90170
#>           [,29]     [,30]
#>  [1,] 331.06193 351.29190
#>  [2,] 328.29560 348.48816
#>  [3,] 325.49194 345.11882
#>  [4,] 324.61670 344.45609
#>  [5,] 319.93281 339.66307
#>  [6,] 299.64479 320.38258
#>  [7,] 290.63723 310.62518
#>  [8,] 279.18632 299.57804
#>  [9,] 272.21499 293.07508
#> [10,] 261.03257 281.59190
#> [11,] 259.01930 279.46556
#> [12,] 257.95736 277.67967
#> [13,] 252.07340 272.22968
#> [14,] 248.06451 268.86800
#> [15,] 245.73563 265.28475
#> [16,] 242.21065 262.95627
#> [17,] 232.96781 252.06348
#> [18,] 228.14469 247.30952
#> [19,] 205.23889 226.41776
#> [20,] 192.31745 211.40956
#> [21,] 189.53100 209.61870
#> [22,] 176.28386 197.71191
#> [23,] 146.42746 166.13549
#> [24,] 141.79210 161.75599
#> [25,]  82.07923  99.69453
#> [26,]  70.63993  92.40130
#> [27,]  36.45545  55.45268
#> [28,]  35.05710  55.90170
#> [29,]   0.00000  26.26785
#> [30,]  26.26785   0.00000