Calculate pairwise distances between two matrices or data frames.

This function takes two data frames or matrices and returns a matrix of pairwise Euclidean distances between the two.

Usage

pairDist(x1, x2 = NULL, na.rm = FALSE)

Arguments

x1

Data frame or matrix one or more columns wide.

x2

Data frame or matrix one or more columns wide. If NULL, then pairwise distances between all points in x1 are calculated.

na.rm

Logical, if TRUE then any rows in x1 or x2 with at least one NA are removed first.

Value

Matrix with nrow(x1) rows and nrow(x2) columns. Values are the distance between each row of x1 and row of x2.

See also

Examples

x1 <- data.frame(x=sample(1:30, 30), y=sort(round(100 * rnorm(30))), z=sample(1:30, 30))
x2 <- data.frame(x=1:20, y=round(100 * rnorm(20)), z=sample(1:20, 20))
pairDist(x1, x2)
#>           [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 449.1392 190.59381 312.40999  56.15158 239.21747 138.20999  40.01250
#>  [2,] 404.6356 147.11220 268.16786  19.51922 194.93845  94.41398  25.35744
#>  [3,] 399.0138 142.42893 262.88020  21.79449 189.68658  89.78864  33.95585
#>  [4,] 367.5037 109.49429 230.63608  32.12476 157.76565  58.14637  47.89572
#>  [5,] 365.5475 108.68763 229.37524  31.01612 156.21140  56.54202  54.91812
#>  [6,] 323.2785  66.15890 186.81809  70.24244 113.55175  14.45683  91.78235
#>  [7,] 314.3644  57.07013 177.61475  80.06248 104.58489  10.63015  99.72963
#>  [8,] 313.6782  58.82176 177.58941  81.12336 104.47488  14.56022 102.63040
#>  [9,] 306.2336  53.60037 170.14406  91.23596  97.82638  22.82542 110.57124
#> [10,] 302.0828  43.70355 165.18475  91.28527  92.04890   9.69536 111.22500
#> [11,] 297.6054  48.85693 162.11416 100.34939  90.08885  28.89637 120.55704
#> [12,] 296.8046  42.10701 160.13432  99.50879  87.94316  24.10394 117.95762
#> [13,] 293.1706  35.39774 156.28500 100.48881  83.22259  19.10497 120.21647
#> [14,] 285.4856  31.84337 149.14423 108.57716  76.08548  27.96426 129.12397
#> [15,] 283.1572  25.01999 146.12324 110.76552  73.35530  29.56349 130.01923
#> [16,] 273.3295  20.42058 136.81009 120.34118  63.70243  38.69108 140.64139
#> [17,] 262.1793  27.53180 127.07478 133.18033  55.39856  53.34791 154.36645
#> [18,] 245.9797  36.52396 112.09817 150.98675  43.78356  71.42829 171.98546
#> [19,] 240.3560  35.05710 106.15555 155.03548  37.06751  74.70609 176.47096
#> [20,] 225.3220  36.93237  89.23004 168.11008  17.29162  86.18585 188.54442
#> [21,] 220.3475  46.82948  85.37564 175.28833  22.13594  93.94147 195.08972
#> [22,] 219.8750  50.80354  86.31918 176.51629  25.96151  95.65041 196.94162
#> [23,] 217.1198  47.94789  81.59044 178.19652  19.07878  96.63850 197.71191
#> [24,] 200.0250  63.32456  65.05382 194.78450  19.51922 113.00442 214.64622
#> [25,] 177.1638  82.75869  41.77320 216.00000  33.55592 134.04850 236.37682
#> [26,] 176.0568  83.07828  39.16631 217.18886  34.13210 135.13697 237.04219
#> [27,] 158.9906 103.23759  30.56141 235.22117  54.41507 153.55129 256.03711
#> [28,] 151.3407 115.88788  35.01428 246.32296  67.94851 164.94241 266.83516
#> [29,] 143.0280 122.24974  27.18455 254.57023  74.49832 172.86122 273.97445
#> [30,] 115.4383 144.67895  26.64583 278.02338  95.47251 196.13261 298.47110
#>            [,8]      [,9]      [,10]     [,11]     [,12]     [,13]      [,14]
#>  [1,] 287.38476 167.10775 145.389133  93.34881 294.15302 107.58253 131.464824
#>  [2,] 242.87445 122.39690 101.034648  48.60041 249.17062  62.84903  86.238043
#>  [3,] 237.57525 116.94871  96.322375  43.90900 243.50359  58.54912  81.049368
#>  [4,] 205.35092  86.84469  64.070274  20.66398 212.43587  27.47726  51.048996
#>  [5,] 204.41869  83.86895  63.960926  15.65248 210.63001  29.47881  49.779514
#>  [6,] 161.70034  41.06093  21.236761  33.24154 168.12198  20.80865   8.831761
#>  [7,] 152.29248  33.97058  11.704700  43.46263 159.16344  28.44293   9.165151
#>  [8,] 152.30890  31.79623  16.309506  43.28972 158.21504  30.69202   7.810250
#>  [9,] 144.61328  30.43025  18.138357  54.21254 150.88075  39.93745  19.748418
#> [10,] 140.05713  22.38303   3.464102  55.19058 147.22092  40.62019  19.313208
#> [11,] 136.57599  24.49490  22.912878  62.64982 142.10911  49.20366  26.795522
#> [12,] 134.76275  25.57342  17.000000  63.60031 142.04577  48.38388  28.354894
#> [13,] 131.06868  15.68439  11.445523  64.13267 138.23531  49.40648  27.092434
#> [14,] 123.78207   7.00000  20.542639  71.17584 130.03846  57.11392  33.030289
#> [15,] 121.02066  16.03122  22.494444  74.98000 128.69343  60.14981  38.405729
#> [16,] 111.47645  11.04536  31.288976  83.24062 118.02118  69.00000  45.188494
#> [17,] 101.81356  23.85372  46.411206  95.44632 106.68646  82.13404  57.784081
#> [18,]  86.92526  42.05948  64.187226 113.11057  90.74139  99.74969  75.478474
#> [19,]  81.36953  44.74371  68.036755 117.52021  85.32878 104.39828  80.087452
#> [20,]  63.98437  57.08765  79.158070 131.06487  70.02857 117.05554  93.107465
#> [21,]  59.89157  65.37584  86.319175 137.94564  65.42935 123.60016  99.789779
#> [22,]  61.22908  66.28725  88.164619 138.82723  64.81512 124.93598 100.796825
#> [23,]  56.08030  68.52737  89.044932 141.11343  62.63386 126.58989 102.985436
#> [24,]  39.67367  84.24369 105.574618 157.57538  45.18849 143.25502 119.411055
#> [25,]  18.46619 105.19506 127.204560 179.20658  23.43075 165.24830 141.385996
#> [26,]  14.35270 106.86908 128.195164 180.64329  24.28992 166.38510 142.790056
#> [27,]  20.49390 124.18132 146.724231 198.12622  11.22497 184.46138 160.352736
#> [28,]  31.59114 135.38464 157.638828 208.70074  19.33908 194.87432 170.683918
#> [29,]  31.12876 144.32602 165.363237 217.42355  26.32489 202.96059 179.265167
#> [30,]  49.13247 167.14963 189.298706 241.17628  41.27953 227.30156 203.354370
#>           [,15]      [,16]     [,17]      [,18]     [,19]      [,20]
#>  [1,] 204.44070  85.475143  21.35416 240.819434 264.75649 231.633331
#>  [2,] 159.34240  41.158231  61.30253 195.739623 219.58597 186.021504
#>  [3,] 153.99351  38.704005  68.27884 190.520340 214.23819 180.227634
#>  [4,] 122.36421   6.708204  98.77753 158.015822 182.01374 149.856598
#>  [5,] 121.74153  23.430749 101.86756 158.458196 182.11260 148.388005
#>  [6,]  78.72103  44.102154 142.38329 115.381974 139.12944 105.806427
#>  [7,]  69.07243  51.000000 151.16547 105.095195 129.06975  96.710909
#>  [8,]  68.89122  53.888774 152.19067 105.621967 129.22074  95.236548
#>  [9,]  61.39218  61.261734 160.42755  96.772930 120.48651  87.732548
#> [10,]  57.56735  63.513778 163.44418  93.584187 117.53297  85.697141
#> [11,]  53.48832  71.070388 169.47271  88.966286 112.32542  78.517514
#> [12,]  52.64029  69.720872 169.73803  87.183714 111.15305  80.305666
#> [13,]  48.42520  72.201108 172.33978  84.314886 108.30051  76.648549
#> [14,]  40.19950  80.311892 180.01111  76.850504 100.57833  67.305275
#> [15,]  39.74921  82.595399 182.71289  74.605630  98.59006  68.337398
#> [16,]  28.08914  92.135769 192.06509  64.691576  88.48164  55.883808
#> [17,]  21.35416 105.385008 204.24495  56.089215  78.61298  43.520110
#> [18,]  18.38478 122.743635 221.43622  42.308392  63.16645  27.495454
#> [19,]  19.20937 127.800626 226.48841  39.509493  59.50630  24.596748
#> [20,]  20.83267 140.256907 240.10414  20.493902  41.97618  13.190906
#> [21,]  27.87472 146.290806 246.18083  13.601471  34.88553  10.344080
#> [22,]  30.51229 147.827602 247.21448  20.024984  36.91883   6.082763
#> [23,]  30.67572 149.194504 249.27294   8.660254  31.25700  14.456832
#> [24,]  46.43275 166.078295 266.09773  11.575837  14.69694  21.494185
#> [25,]  68.91299 188.448932 288.30019  36.069378  18.78829  44.418465
#> [26,]  70.07139 189.362087 289.45811  35.468296  16.67333  47.053161
#> [27,]  88.28363 207.783541 307.30116  55.731499  34.95712  61.741396
#> [28,]  98.60020 217.866932 317.28536  64.822835  42.35564  70.526591
#> [29,] 106.44717 225.514966 325.51958  70.064256  46.42198  80.454956
#> [30,] 130.74402 250.519460 350.32271  96.104110  72.86975 104.604015
pairDist(x1)
#>            [,1]       [,2]       [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,]   0.00000  47.602521  54.571055  84.90583  84.50444 126.06744 135.92277
#>  [2,]  47.60252   0.000000   9.273618  41.19466  42.83690  81.79853  90.95603
#>  [3,]  54.57105   9.273618   0.000000  39.83717  37.64306  76.60940  86.40023
#>  [4,]  84.90583  41.194660  39.837169   0.00000  29.89983  48.27007  53.38539
#>  [5,]  84.50444  42.836900  37.643060  29.89983   0.00000  43.17407  56.12486
#>  [6,] 126.06744  81.798533  76.609399  48.27007  43.17407   0.00000  16.18641
#>  [7,] 135.92277  90.956033  86.400231  53.38539  56.12486  16.18641   0.00000
#>  [8,] 136.53571  91.093359  85.358069  57.55867  53.74942  12.84523  14.17745
#>  [9,] 146.27713 100.005000  94.968416  63.02381  67.48333  28.53069  15.29706
#> [10,] 147.42456 103.440804  98.883770  66.21933  66.17401  23.72762  14.45683
#> [11,] 155.20631 108.577162 102.883429  73.33485  74.79305  34.84250  24.77902
#> [12,] 155.10964 110.095413 105.815878  71.09149  76.66812  35.74913  20.59126
#> [13,] 156.59183 112.227448 107.577879  74.60563  75.39231  32.55764  21.67948
#> [14,] 164.45060 119.188926 113.815640  83.02409  81.95731  38.94868  29.84962
#> [15,] 166.87720 122.971541 118.549568  84.74078  86.09878  43.50862  32.54228
#> [16,] 176.35476 131.403957 126.162594  94.68896  93.82963  50.71489  41.30375
#> [17,] 188.69022 143.059428 137.047437 108.30051 105.22832  63.19019  55.66866
#> [18,] 206.31529 160.224842 154.084392 125.38341 123.17061  81.30806  72.97260
#> [19,] 210.53503 165.302753 159.157155 130.85870 126.51482  84.86460  77.98718
#> [20,] 224.12943 179.315365 173.798734 142.77605 140.80128  98.12747  89.40358
#> [21,] 231.15579 185.458890 180.002778 148.24979 148.82204 105.84895  95.44632
#> [22,] 232.16158 186.174649 180.313616 150.08664 149.23136 106.72394  97.26253
#> [23,] 234.16447 188.745861 183.491144 151.04304 152.07235 108.96789  98.29547
#> [24,] 250.77480 205.372832 199.914982 168.11008 168.06249 125.11195 115.11299
#> [25,] 272.03125 227.571967 222.015765 190.93454 188.47281 146.11639 137.61541
#> [26,] 273.31667 229.002183 223.731983 191.59071 190.31815 147.67871 138.49549
#> [27,] 291.07731 246.300629 240.403827 210.42101 207.08211 165.07877 157.03821
#> [28,] 302.03642 256.175721 250.159949 220.11588 218.56120 176.24131 167.13168
#> [29,] 310.53985 264.954713 259.539978 227.22236 228.12716 185.13239 174.66539
#> [30,] 334.00299 289.485751 283.763986 252.97628 250.20192 208.06970 199.65220
#>            [,8]      [,9]      [,10]     [,11]     [,12]      [,13]     [,14]
#>  [1,] 136.53571 146.27713 147.424557 155.20631 155.10964 156.591826 164.45060
#>  [2,]  91.09336 100.00500 103.440804 108.57716 110.09541 112.227448 119.18893
#>  [3,]  85.35807  94.96842  98.883770 102.88343 105.81588 107.577879 113.81564
#>  [4,]  57.55867  63.02381  66.219333  73.33485  71.09149  74.605630  83.02409
#>  [5,]  53.74942  67.48333  66.174013  74.79305  76.66812  75.392307  81.95731
#>  [6,]  12.84523  28.53069  23.727621  34.84250  35.74913  32.557641  38.94868
#>  [7,]  14.17745  15.29706  14.456832  24.77902  20.59126  21.679483  29.84962
#>  [8,]   0.00000  19.51922  19.544820  22.86919  28.93095  25.787594  28.87906
#>  [9,]  19.51922   0.00000  20.615528  12.08305  15.42725  21.400935  24.18677
#> [10,]  19.54482  20.61553   0.000000  24.91987  16.76305   9.539392  20.04994
#> [11,]  22.86919  12.08305  24.919872   0.00000  20.24846  22.405357  18.24829
#> [12,]  28.93095  15.42725  16.763055  20.24846   0.00000  12.884099  19.87461
#> [13,]  25.78759  21.40093   9.539392  22.40536  12.88410   0.000000  12.68858
#> [14,]  28.87906  24.18677  20.049938  18.24829  19.87461  12.688578   0.00000
#> [15,]  36.95944  30.74085  19.949937  29.88311  18.41195  11.357817  15.36229
#> [16,]  41.24318  35.35534  29.849623  28.87906  27.45906  20.976177  12.52996
#> [17,]  52.17279  47.46578  45.672749  37.48333  43.30127  37.696154  26.07681
#> [18,]  69.89993  63.19019  63.403470  52.84884  58.62593  54.872580  43.65776
#> [19,]  74.35725  70.25667  66.850580  60.31584  64.73021  58.702640  48.17676
#> [20,]  88.45338  82.05486  77.343390  73.66817  73.35530  68.476273  60.21628
#> [21,]  95.39916  86.05812  84.776176  77.62731  77.70457  75.418831  66.94027
#> [22,]  95.77578  87.48714  86.884981  78.06408  80.62258  77.756029  68.08083
#> [23,]  98.88883  89.34204  87.321246  81.43709  80.16234  77.935871  70.17834
#> [24,] 114.95216 106.24971 103.836410  97.84171  97.42176  94.578010  86.41759
#> [25,] 136.66382 130.24592 125.207827 121.74564 121.05371 116.421647 108.57716
#> [26,] 138.47021 131.22119 126.071408 123.29234 121.26417 117.170815 110.00909
#> [27,] 155.20954 149.12076 144.899965 139.96071 140.81548 136.157996 127.56959
#> [28,] 165.52341 157.54047 156.096124 148.09794 150.01667 146.986394 137.81872
#> [29,] 174.88568 165.25738 163.563443 156.92673 156.21780 154.227105 146.35231
#> [30,] 198.45402 191.83587 187.312573 183.02732 182.80317 178.507703 170.49927
#>           [,15]     [,16]     [,17]     [,18]     [,19]     [,20]      [,21]
#>  [1,] 166.87720 176.35476 188.69022 206.31529 210.53503 224.12943 231.155792
#>  [2,] 122.97154 131.40396 143.05943 160.22484 165.30275 179.31536 185.458890
#>  [3,] 118.54957 126.16259 137.04744 154.08439 159.15716 173.79873 180.002778
#>  [4,]  84.74078  94.68896 108.30051 125.38341 130.85870 142.77605 148.249789
#>  [5,]  86.09878  93.82963 105.22832 123.17061 126.51482 140.80128 148.822041
#>  [6,]  43.50862  50.71489  63.19019  81.30806  84.86460  98.12747 105.848949
#>  [7,]  32.54228  41.30375  55.66866  72.97260  77.98718  89.40358  95.446320
#>  [8,]  36.95944  41.24318  52.17279  69.89993  74.35725  88.45338  95.399161
#>  [9,]  30.74085  35.35534  47.46578  63.19019  70.25667  82.05486  86.058120
#> [10,]  19.94994  29.84962  45.67275  63.40347  66.85058  77.34339  84.776176
#> [11,]  29.88311  28.87906  37.48333  52.84884  60.31584  73.66817  77.627315
#> [12,]  18.41195  27.45906  43.30127  58.62593  64.73021  73.35530  77.704569
#> [13,]  11.35782  20.97618  37.69615  54.87258  58.70264  68.47627  75.418831
#> [14,]  15.36229  12.52996  26.07681  43.65776  48.17676  60.21628  66.940272
#> [15,]   0.00000  15.39480  33.52611  49.21382  52.23983  59.44746  66.580778
#> [16,]  15.39480   0.00000  18.13836  34.45287  38.02631  48.13523  55.353410
#> [17,]  33.52611  18.13836   0.00000  18.49324  22.91288  38.28838  45.310043
#> [18,]  49.21382  34.45287  18.49324   0.00000  13.15295  26.94439  30.248967
#> [19,]  52.23983  38.02631  22.91288  13.15295   0.00000  20.61553  30.495901
#> [20,]  59.44746  48.13523  38.28838  26.94439  20.61553   0.00000  16.643317
#> [21,]  66.58078  55.35341  45.31004  30.24897  30.49590  16.64332   0.000000
#> [22,]  69.82120  57.07889  44.21538  27.00000  26.79552  18.46619  10.099505
#> [23,]  68.56384  58.29237  49.60847  35.62303  35.01428  18.19341   6.164414
#> [24,]  85.22910  74.49161  64.15606  48.55924  45.61798  28.67054  20.420578
#> [25,] 106.79419  96.42614  85.84288  71.30919  64.52906  48.38388  47.053161
#> [26,] 107.11676  97.65756  88.37420  74.25631  68.13956  50.25933  47.801674
#> [27,] 126.95669 115.73677 103.60502  87.95453  81.44323  67.72001  65.199693
#> [28,] 138.12313 126.35268 113.44602  96.09370  92.36341  79.48585  72.394751
#> [29,] 144.53719 134.45445 123.85072 107.13076 104.18253  88.14193  79.561297
#> [30,] 168.81351 158.43295 147.04421 131.32403 125.07997 110.30866 106.681770
#>           [,22]      [,23]     [,24]      [,25]      [,26]     [,27]     [,28]
#>  [1,] 232.16158 234.164472 250.77480 272.031248 273.316666 291.07731 302.03642
#>  [2,] 186.17465 188.745861 205.37283 227.571967 229.002183 246.30063 256.17572
#>  [3,] 180.31362 183.491144 199.91498 222.015765 223.731983 240.40383 250.15995
#>  [4,] 150.08664 151.043040 168.11008 190.934544 191.590710 210.42101 220.11588
#>  [5,] 149.23136 152.072351 168.06249 188.472810 190.318155 207.08211 218.56120
#>  [6,] 106.72394 108.967885 125.11195 146.116392 147.678705 165.07877 176.24131
#>  [7,]  97.26253  98.295473 115.11299 137.615406 138.495487 157.03821 167.13168
#>  [8,]  95.77578  98.888826 114.95216 136.663821 138.470213 155.20954 165.52341
#>  [9,]  87.48714  89.342039 106.24971 130.245921 131.221187 149.12076 157.54047
#> [10,]  86.88498  87.321246 103.83641 125.207827 126.071408 144.89997 156.09612
#> [11,]  78.06408  81.437092  97.84171 121.745636 123.292336 139.96071 148.09794
#> [12,]  80.62258  80.162335  97.42176 121.053707 121.264174 140.81548 150.01667
#> [13,]  77.75603  77.935871  94.57801 116.421647 117.170815 136.15800 146.98639
#> [14,]  68.08083  70.178344  86.41759 108.577162 110.009091 127.56959 137.81872
#> [15,]  69.82120  68.563839  85.22910 106.794195 107.116759 126.95669 138.12313
#> [16,]  57.07889  58.292367  74.49161  96.426138  97.657565 115.73677 126.35268
#> [17,]  44.21538  49.608467  64.15606  85.842880  88.374204 103.60502 113.44602
#> [18,]  27.00000  35.623026  48.55924  71.309186  74.256313  87.95453  96.09370
#> [19,]  26.79552  35.014283  45.61798  64.529063  68.139563  81.44323  92.36341
#> [20,]  18.46619  18.193405  28.67054  48.383882  50.259327  67.72001  79.48585
#> [21,]  10.09950   6.164414  20.42058  47.053161  47.801674  65.19969  72.39475
#> [22,]   0.00000  15.427249  22.51666  47.053161  49.608467  63.31666  70.12132
#> [23,]  15.42725   0.000000  17.57840  44.113490  43.874822  63.06346  70.71775
#> [24,]  22.51666  17.578396   0.00000  28.089144  28.460499  45.84757  53.49766
#> [25,]  47.05316  44.113490  28.08914   0.000000   9.110434  21.56386  38.61347
#> [26,]  49.60847  43.874822  28.46050   9.110434   0.000000  26.19160  40.84116
#> [27,]  63.31666  63.063460  45.84757  21.563859  26.191602   0.00000  23.28089
#> [28,]  70.12132  70.717749  53.49766  38.613469  40.841156  23.28089   0.00000
#> [29,]  80.14986  76.380626  60.09160  46.540305  43.874822  36.61967  23.60085
#> [30,] 105.83478 103.744879  86.40602  62.104750  62.241465  43.74929  42.16634
#>           [,29]     [,30]
#>  [1,] 310.53985 334.00299
#>  [2,] 264.95471 289.48575
#>  [3,] 259.53998 283.76399
#>  [4,] 227.22236 252.97628
#>  [5,] 228.12716 250.20192
#>  [6,] 185.13239 208.06970
#>  [7,] 174.66539 199.65220
#>  [8,] 174.88568 198.45402
#>  [9,] 165.25738 191.83587
#> [10,] 163.56344 187.31257
#> [11,] 156.92673 183.02732
#> [12,] 156.21780 182.80317
#> [13,] 154.22711 178.50770
#> [14,] 146.35231 170.49927
#> [15,] 144.53719 168.81351
#> [16,] 134.45445 158.43295
#> [17,] 123.85072 147.04421
#> [18,] 107.13076 131.32403
#> [19,] 104.18253 125.07997
#> [20,]  88.14193 110.30866
#> [21,]  79.56130 106.68177
#> [22,]  80.14986 105.83478
#> [23,]  76.38063 103.74488
#> [24,]  60.09160  86.40602
#> [25,]  46.54031  62.10475
#> [26,]  43.87482  62.24147
#> [27,]  36.61967  43.74929
#> [28,]  23.60085  42.16634
#> [29,]   0.00000  40.06245
#> [30,]  40.06245   0.00000