Calculate pairwise distances between two matrices or data frames.

This function takes two data frames or matrices and returns a matrix of pairwise Euclidean distances between the two.

Usage

pairDist(x1, x2 = NULL, na.rm = FALSE)

Arguments

x1

Data frame or matrix one or more columns wide.

x2

Data frame or matrix one or more columns wide. If NULL, then pairwise distances between all points in x1 are calculated.

na.rm

Logical, if TRUE then any rows in x1 or x2 with at least one NA are removed first.

Value

Matrix with nrow(x1) rows and nrow(x2) columns. Values are the distance between each row of x1 and row of x2.

See also

Examples

x1 <- data.frame(x=sample(1:30, 30), y=sort(round(100 * rnorm(30))), z=sample(1:30, 30))
x2 <- data.frame(x=1:20, y=round(100 * rnorm(20)), z=sample(1:20, 20))
pairDist(x1, x2)
#>            [,1]      [,2]      [,3]      [,4]       [,5]      [,6]      [,7]
#>  [1,] 141.17719 145.71548 253.80898 285.51883 169.994118  21.74856  44.65423
#>  [2,] 132.61222 138.28955 246.15036 278.07193 162.126494  13.41641  45.70558
#>  [3,]  90.04999  96.33795 202.72395 234.59966 118.987394  55.79426  89.09545
#>  [4,]  76.38063  82.82512 190.06841 222.07206 106.061303  66.06058 100.18483
#>  [5,]  72.94518  79.15175 186.21761 218.16737 102.244804  70.32069 104.35516
#>  [6,]  69.86415  74.06079 182.39792 214.19851  98.661036  75.61085 109.77249
#>  [7,]  45.40925  53.12250 159.12888 191.19100  75.266194  97.02577 131.01527
#>  [8,]  39.03844  39.16631 147.11900 178.63370  64.288413 111.99107 146.16771
#>  [9,]  31.71750  35.91657 138.32570 169.95882  55.668663 120.17487 153.99675
#> [10,]  33.79349  33.49627 138.52798 169.93528  56.098128 120.80563 154.82571
#> [11,]  27.00000  31.92178 123.50304 154.98064  43.611925 136.65650 170.18226
#> [12,]  29.58040  21.47091 121.53600 152.54180  41.785165 139.44533 173.51081
#> [13,]  29.89983  17.83255 103.46980 134.40982  26.702060 156.81199 190.94764
#> [14,]  33.07567  26.94439 104.52751 135.32923  29.983329 157.09233 190.92145
#> [15,]  27.45906  18.60108  93.96808 125.46713  14.456832 163.59095 197.73973
#> [16,]  37.38984  29.12044  95.71834 126.33289  25.475478 165.98795 199.91248
#> [17,]  26.03843  20.14944  91.40022 123.19903   9.433981 165.18475 199.29124
#> [18,]  42.05948  35.62303  90.49309 120.93387  26.038433 172.03779 205.83974
#> [19,]  42.70831  32.95451  85.46344 115.99138  22.022716 175.67584 209.73555
#> [20,]  38.00000  34.55431  76.13147 108.26819  10.246951 180.07221 214.10511
#> [21,]  59.24525  52.19195  56.51548  88.11356  28.653098 200.24984 234.42056
#> [22,]  74.49161  66.27971  44.59821  74.97333  44.317040 214.77663 249.00602
#> [23,]  84.08329  78.29432  38.60052  67.74954  54.083269 223.68281 257.62182
#> [24,]  85.79627  81.50460  42.76681  70.50532  57.078893 224.23871 257.94185
#> [25,]  85.42833  79.18333  29.29164  61.00820  55.090834 227.11671 261.24892
#> [26,]  89.82761  84.59905  30.67572  60.20797  59.799666 230.36710 264.30286
#> [27,] 110.46719 103.38762  24.71841  44.55334  79.937476 250.04200 284.08097
#> [28,] 121.09913 116.09048  16.76305  30.74085  91.148231 262.26132 296.16887
#> [29,] 153.22206 146.84345  43.46263  21.79449 122.743635 293.75330 327.75296
#> [30,] 153.10454 145.58503  42.48529  18.41195 122.547950 293.91665 328.11888
#>             [,8]      [,9]      [,10]     [,11]     [,12]     [,13]     [,14]
#>  [1,] 169.345800  19.94994 196.613835 317.99843 137.94564 316.26887  59.30430
#>  [2,] 162.040118  12.12436 189.023808 310.27407 129.48359 309.01618  52.49762
#>  [3,] 119.029408  36.79674 145.399450 266.21983  85.46929 265.16033  16.40122
#>  [4,] 106.381389  46.33573 133.048863 254.11415  73.28028 253.09879  14.00000
#>  [5,] 102.391406  50.09990 129.031004 250.12997  69.26038 249.03413  13.74773
#>  [6,]  98.264948  53.49766 125.654288 246.97571  68.18358 245.37726  17.00000
#>  [7,]  76.124897  77.49194 102.283919 223.05605  42.30839 222.26336  39.11521
#>  [8,]  62.872888  89.44831  90.564894 211.68845  38.28838 209.62109  49.33559
#>  [9,]  55.145263  99.18165  80.894994 201.54900  23.74868 200.21239  58.25805
#> [10,]  54.267854  98.49365  81.289606 202.32153  28.91366 200.31226  57.21888
#> [11,]  43.886217 116.51609  66.700825 185.70137  15.52417 184.80260  76.07233
#> [12,]  38.183766 116.73046  65.122961 185.35102  27.36786 182.81685  75.69016
#> [13,]  21.424285 134.43586  47.434165 167.24234  29.83287 164.69669  93.34881
#> [14,]  26.324893 135.34031  48.207883 166.84724  27.71281 164.79381  94.05318
#> [15,]  10.344080 142.05985  36.796739 157.93036  28.31960 156.08011 101.02475
#> [16,]  20.322401 144.03819  40.024992 158.12021  34.38023 155.80436 102.72293
#> [17,]   9.110434 144.18391  34.029399 155.38983  27.53180 154.01623 103.35376
#> [18,]  22.090722 150.48588  35.735137 152.04276  38.48376 150.02000 109.21996
#> [19,]  15.394804 153.58711  30.692019 148.29700  42.30839 145.71548 112.32987
#> [20,]  16.031220 159.81552  21.863211 140.41367  42.49706 139.74620 119.65785
#> [21,]  28.089144 179.05027   6.480741 121.03718  62.65780 119.24345 138.28955
#> [22,]  42.871902 193.15797  21.047565 108.76121  78.40281 106.04716 152.44999
#> [23,]  53.619026 202.68448  27.313001  98.15294  85.00588  96.56604 161.56113
#> [24,]  57.428216 203.78420  32.202484  98.65597  85.72631  98.06121 162.80970
#> [25,]  55.290144 206.14801  28.705400  94.03191  88.92694  92.43917 165.45090
#> [26,]  60.049979 209.64255  32.664966  90.49309  91.31813  89.47625 168.65646
#> [27,]  78.701969 228.49508  52.697249  74.63913 111.95088  71.20393 187.26719
#> [28,]  91.684241 241.73126  64.023433  58.31809 123.10565  58.18935 200.83077
#> [29,] 122.204746 272.49954  95.273291  33.48134 155.01290  28.70540 231.31580
#> [30,] 121.597697 272.18376  95.713113  38.94868 156.20819  30.78961 231.22932
#>           [,15]      [,16]     [,17]      [,18]     [,19]     [,20]
#>  [1,] 123.40989  57.419509 208.79895 267.705062 209.40392 205.38257
#>  [2,] 116.12493  47.455242 201.21133 260.217217 202.20040 198.25489
#>  [3,]  72.73239   5.385165 157.03184 216.009259 158.15815 154.20765
#>  [4,]  60.93439  13.416408 145.27904 204.247399 146.53668 142.67796
#>  [5,]  56.51548  15.394804 141.09217 200.099975 142.26032 138.36184
#>  [6,]  53.83308  28.442925 138.43049 197.149689 139.21925 135.33662
#>  [7,]  32.38827  41.976184 114.50764 173.375315 116.05602 112.29871
#>  [8,]  21.84033  59.715995 103.46980 161.768353 103.77379  99.86491
#>  [9,]  11.09054  63.671030  92.18460 151.105923  93.09672  89.11229
#> [10,]  10.34408  65.833122  93.22017 151.845316  93.43982  89.38121
#> [11,]  19.84943  79.768415  76.81797 135.358783  78.19207  74.31016
#> [12,]  17.83255  85.070559  77.20751 134.988888  76.74634  72.67737
#> [13,]  31.38471 101.872469  59.49790 116.940156  58.83026  54.79964
#> [14,]  31.92178 100.846418  58.12056 116.056021  57.92236  53.74942
#> [15,]  37.10795 107.953694  49.19350 107.680082  49.52777  45.63989
#> [16,]  40.01250 109.945441  49.85980 107.354553  49.16299  44.95553
#> [17,]  39.40812 109.407495  46.56179 105.351792  47.71792  44.03408
#> [18,]  46.49731 115.589792  43.94315 101.212647  43.52011  39.34463
#> [19,]  49.20366 119.958326  40.91455  97.718985  39.61060  35.45420
#> [20,]  56.85948 124.907966  35.01428  91.613318  37.80212  35.28456
#> [21,]  74.54529 144.844744  17.49286  71.589105  17.80449  16.15549
#> [22,]  89.17399 159.790488  21.47091  60.415230  18.97367  20.14944
#> [23,]  97.51410 167.254297  13.56466  47.360321  12.36932  16.09348
#> [24,]  99.14636 167.532086  18.35756  48.641546  19.74842  22.40536
#> [25,] 101.64153 171.656634  21.21320  45.617979  21.23676  24.91987
#> [26,] 104.55620 174.037352  19.20937  40.024992  19.69772  23.79075
#> [27,] 123.39368 193.827243  40.45986  25.416530  37.81534  41.49699
#> [28,] 136.73332 206.012136  51.35173   9.486833  51.45872  55.62374
#> [29,] 167.30212 237.425778  82.68615  25.337719  81.22807  85.14693
#> [30,] 167.61265 238.417701  85.02353  33.075671  82.98193  86.84469
pairDist(x1)
#>            [,1]      [,2]      [,3]       [,4]       [,5]      [,6]      [,7]
#>  [1,]   0.00000  14.45683  54.75400  65.795137  69.152006  72.18033  96.79359
#>  [2,]  14.45683   0.00000  45.66180  56.356011  60.241182  64.87681  87.62420
#>  [3,]  54.75400  45.66180   0.00000  17.000000  18.000000  30.33150  45.02222
#>  [4,]  65.79514  56.35601  17.00000   0.000000   5.744563  16.40122  31.27299
#>  [5,]  69.15201  60.24118  18.00000   5.744563   0.000000  16.12452  27.91057
#>  [6,]  72.18033  64.87681  30.33150  16.401219  16.124515   0.00000  29.03446
#>  [7,]  96.79359  87.62420  45.02222  31.272992  27.910571  29.03446   0.00000
#>  [8,] 107.38249 100.96534  61.32699  48.020829  44.418465  36.61967  26.49528
#>  [9,] 117.23907 109.69959  65.15366  54.899909  50.049975  50.24938  27.53180
#> [10,] 116.10771 109.66312  66.97014  55.982140  51.312766  47.80167  31.08054
#> [11,] 134.61798 126.62938  81.40025  71.742595  66.977608  68.33740  42.95346
#> [12,] 134.07461 128.14445  86.22065  74.759615  70.327804  65.16134  48.54894
#> [13,] 152.01645 145.76008 103.28117  91.487704  87.149297  82.33468  63.27717
#> [14,] 152.80380 146.19850 102.02451  91.880357  87.068938  85.09407  63.48228
#> [15,] 160.20300 153.08168 109.77705  97.642204  93.439820  89.72736  67.73478
#> [16,] 161.49613 155.03548 111.18003 100.618090  95.921843  93.15042  71.93052
#> [17,] 162.63149 155.01613 111.47645  99.126182  95.052617  91.88580  68.58571
#> [18,] 168.00595 161.27306 116.85889 106.681770 101.921538 100.07997  77.63376
#> [19,] 171.14614 164.73615 121.34661 110.099955 105.607765 101.86756  81.03086
#> [20,] 178.71486 170.47580 127.44018 114.254103 110.620975 107.20541  83.22259
#> [21,] 197.41580 190.05262 146.94557 134.253492 130.326513 126.16259 103.70149
#> [22,] 211.29600 204.36487 161.79617 148.946299 145.058609 140.03571 118.76447
#> [23,] 220.76458 213.43383 169.05325 157.727613 153.398175 150.87412 127.15345
#> [24,] 221.92341 214.24052 169.34580 158.552830 154.149278 152.74161 127.94139
#> [25,] 224.59964 217.08524 173.84188 161.127279 157.254571 153.22206 130.34569
#> [26,] 227.93420 220.32930 176.00568 164.380656 160.181147 157.63883 133.55523
#> [27,] 246.27627 239.48486 195.47378 184.089652 179.738699 176.38027 153.78231
#> [28,] 260.12689 252.38265 208.08171 196.298242 192.192612 189.55210 165.31485
#> [29,] 290.41694 283.39901 239.18612 227.743277 223.459169 220.28618 197.13194
#> [30,] 290.05861 283.39901 240.26236 227.971489 223.915163 219.25784 197.59808
#>            [,8]      [,9]     [,10]     [,11]     [,12]     [,13]     [,14]
#>  [1,] 107.38249 117.23907 116.10771 134.61798 134.07461 152.01645 152.80380
#>  [2,] 100.96534 109.69959 109.66312 126.62938 128.14445 145.76008 146.19850
#>  [3,]  61.32699  65.15366  66.97014  81.40025  86.22065 103.28117 102.02451
#>  [4,]  48.02083  54.89991  55.98214  71.74260  74.75961  91.48770  91.88036
#>  [5,]  44.41846  50.04998  51.31277  66.97761  70.32780  87.14930  87.06894
#>  [6,]  36.61967  50.24938  47.80167  68.33740  65.16134  82.33468  85.09407
#>  [7,]  26.49528  27.53180  31.08054  42.95346  48.54894  63.27717  63.48228
#>  [8,]   0.00000  25.01999  16.30951  41.04875  29.30870  46.10857  51.16640
#>  [9,]  25.01999   0.00000  12.64911  18.57418  26.85144  40.42277  37.36308
#> [10,]  16.30951  12.64911   0.00000  27.14774  19.41649  36.46917  37.57659
#> [11,]  41.04875  18.57418  27.14774   0.00000  29.22328  34.19064  25.63201
#> [12,]  29.30870  26.85144  19.41649  29.22328   0.00000  18.13836  24.26932
#> [13,]  46.10857  40.42277  36.46917  34.19064  18.13836   0.00000  15.81139
#> [14,]  51.16640  37.36308  37.57659  25.63201  24.26932  15.81139   0.00000
#> [15,]  54.35071  45.45327  44.56456  34.71311  28.79236  12.96148  17.02939
#> [16,]  58.15497  46.49731  45.71652  34.68429  29.88311  15.42725   9.69536
#> [17,]  57.32364  47.30750  47.81213  35.17101  33.66007  19.28730  20.92845
#> [18,]  65.76473  51.97115  52.73519  37.81534  38.13135  23.89561  15.26434
#> [19,]  66.00000  56.76266  55.11806  45.17743  37.66962  20.14944  21.11871
#> [20,]  73.48469  64.79197  65.83312  52.00961  51.95190  36.74235  39.37004
#> [21,]  90.67524  82.71638  82.18272  69.53416  65.30697  47.43416  50.17968
#> [22,] 104.06248  97.74968  96.18212  85.19390  78.31986  60.34070  64.27286
#> [23,] 115.68059 104.00961 104.92855  88.36855  88.57200  70.79548  68.77500
#> [24,] 118.41030 104.71390 106.86908  88.04544  91.68424  74.49161  70.34913
#> [25,] 117.83887 109.50799 109.36178  95.35722  92.37424  74.36397  76.13147
#> [26,] 122.62545 111.15305 112.29871  95.38344  96.09370  78.26238  76.64855
#> [27,] 140.46708 130.32651 130.17296 115.29961 112.72977  94.76814  94.09038
#> [28,] 154.50890 143.37015 144.53719 127.39702 128.10152 110.13174 108.89904
#> [29,] 184.48035 174.07182 174.32441 158.46766 156.97771 138.97122 138.09055
#> [30,] 183.05464 175.41095 174.33588 161.13349 156.16017 138.13399 139.66030
#>           [,15]      [,16]     [,17]      [,18]     [,19]     [,20]     [,21]
#>  [1,] 160.20300 161.496130 162.63149 168.005952 171.14614 178.71486 197.41580
#>  [2,] 153.08168 155.035480 155.01613 161.273060 164.73615 170.47580 190.05262
#>  [3,] 109.77705 111.180034 111.47645 116.858889 121.34661 127.44018 146.94557
#>  [4,]  97.64220 100.618090  99.12618 106.681770 110.09995 114.25410 134.25349
#>  [5,]  93.43982  95.921843  95.05262 101.921538 105.60776 110.62097 130.32651
#>  [6,]  89.72736  93.150416  91.88580 100.079968 101.86756 107.20541 126.16259
#>  [7,]  67.73478  71.930522  68.58571  77.633756  81.03086  83.22259 103.70149
#>  [8,]  54.35071  58.154965  57.32364  65.764732  66.00000  73.48469  90.67524
#>  [9,]  45.45327  46.497312  47.30750  51.971146  56.76266  64.79197  82.71638
#> [10,]  44.56456  45.716518  47.81213  52.735187  55.11806  65.83312  82.18272
#> [11,]  34.71311  34.684290  35.17101  37.815341  45.17743  52.00961  69.53416
#> [12,]  28.79236  29.883106  33.66007  38.131352  37.66962  51.95190  65.30697
#> [13,]  12.96148  15.427249  19.28730  23.895606  20.14944  36.74235  47.43416
#> [14,]  17.02939   9.695360  20.92845  15.264338  21.11871  39.37004  50.17968
#> [15,]   0.00000  13.638182   6.63325  18.411953  15.29706  24.53569  37.70942
#> [16,]  13.63818   0.000000  18.05547   8.774964  11.57584  34.20526  41.61730
#> [17,]   6.63325  18.055470   0.00000  20.273135  18.70829  19.33908  35.58089
#> [18,]  18.41195   8.774964  20.27313   0.000000  12.68858  32.93934  38.43176
#> [19,]  15.29706  11.575837  18.70829  12.688578   0.00000  28.91366  31.33688
#> [20,]  24.53569  34.205263  19.33908  32.939338  28.91366   0.00000  23.36664
#> [21,]  37.70942  41.617304  35.58089  38.431758  31.33688  23.36664   0.00000
#> [22,]  52.31635  55.072679  51.07837  52.096065  43.87482  39.00000  16.40122
#> [23,]  61.35145  60.282667  59.26213  53.916602  51.28353  48.55924  29.22328
#> [24,]  64.38167  62.697687  61.71710  55.371473  55.14526  51.46844  35.81899
#> [25,]  64.82284  67.334983  62.40192  62.777384  56.93856  47.77028  27.23968
#> [26,]  68.33008  68.271517  65.76473  61.983869  59.16925  53.15073  34.19064
#> [27,]  86.72370  85.047046  85.37564  79.284299  75.22633  74.26305  52.35456
#> [28,] 100.36434 100.424101  97.69852  94.116949  91.04395  83.60024  64.50581
#> [29,] 130.58714 129.193653 128.79829 123.117830 119.48640 115.95258  94.92629
#> [30,] 130.52586 130.303492 129.13946 125.355494 119.67038 115.79724  93.95212
#>           [,22]     [,23]     [,24]     [,25]     [,26]     [,27]     [,28]
#>  [1,] 211.29600 220.76458 221.92341 224.59964 227.93420 246.27627 260.12689
#>  [2,] 204.36487 213.43383 214.24052 217.08524 220.32930 239.48486 252.38265
#>  [3,] 161.79617 169.05325 169.34580 173.84188 176.00568 195.47378 208.08171
#>  [4,] 148.94630 157.72761 158.55283 161.12728 164.38066 184.08965 196.29824
#>  [5,] 145.05861 153.39817 154.14928 157.25457 160.18115 179.73870 192.19261
#>  [6,] 140.03571 150.87412 152.74161 153.22206 157.63883 176.38027 189.55210
#>  [7,] 118.76447 127.15345 127.94139 130.34569 133.55523 153.78231 165.31485
#>  [8,] 104.06248 115.68059 118.41030 117.83887 122.62545 140.46708 154.50890
#>  [9,]  97.74968 104.00961 104.71390 109.50799 111.15305 130.32651 143.37015
#> [10,]  96.18212 104.92855 106.86908 109.36178 112.29871 130.17296 144.53719
#> [11,]  85.19390  88.36855  88.04544  95.35722  95.38344 115.29961 127.39702
#> [12,]  78.31986  88.57200  91.68424  92.37424  96.09370 112.72977 128.10152
#> [13,]  60.34070  70.79548  74.49161  74.36397  78.26238  94.76814 110.13174
#> [14,]  64.27286  68.77500  70.34913  76.13147  76.64855  94.09038 108.89904
#> [15,]  52.31635  61.35145  64.38167  64.82284  68.33008  86.72370 100.36434
#> [16,]  55.07268  60.28267  62.69769  67.33498  68.27152  85.04705 100.42410
#> [17,]  51.07837  59.26213  61.71710  62.40192  65.76473  85.37564  97.69852
#> [18,]  52.09607  53.91660  55.37147  62.77738  61.98387  79.28430  94.11695
#> [19,]  43.87482  51.28353  55.14526  56.93856  59.16925  75.22633  91.04395
#> [20,]  39.00000  48.55924  51.46844  47.77028  53.15073  74.26305  83.60024
#> [21,]  16.40122  29.22328  35.81899  27.23968  34.19064  52.35456  64.50581
#> [22,]   0.00000  26.01922  35.86084  17.05872  28.91366  41.01219  54.62600
#> [23,]  26.01922   0.00000  10.81665  19.64688   9.00000  27.87472  40.26164
#> [24,]  35.86084  10.81665   0.00000  28.01785  13.19091  32.58834  40.81666
#> [25,]  17.05872  19.64688  28.01785   0.00000  16.82260  30.54505  38.92300
#> [26,]  28.91366   9.00000  13.19091  16.82260   0.00000  24.12468  32.31099
#> [27,]  41.01219  27.87472  32.58834  30.54505  24.12468   0.00000  23.28089
#> [28,]  54.62600  40.26164  40.81666  38.92300  32.31099  23.28089   0.00000
#> [29,]  82.49848  70.20684  71.49825  69.28925  63.92183  44.29447  34.55431
#> [30,]  79.38514  73.87151  78.00000  68.09552  67.80855  47.34976  41.88078
#>           [,29]     [,30]
#>  [1,] 290.41694 290.05861
#>  [2,] 283.39901 283.39901
#>  [3,] 239.18612 240.26236
#>  [4,] 227.74328 227.97149
#>  [5,] 223.45917 223.91516
#>  [6,] 220.28618 219.25784
#>  [7,] 197.13194 197.59808
#>  [8,] 184.48035 183.05464
#>  [9,] 174.07182 175.41095
#> [10,] 174.32441 174.33588
#> [11,] 158.46766 161.13349
#> [12,] 156.97771 156.16017
#> [13,] 138.97122 138.13399
#> [14,] 138.09055 139.66030
#> [15,] 130.58714 130.52586
#> [16,] 129.19365 130.30349
#> [17,] 128.79829 129.13946
#> [18,] 123.11783 125.35549
#> [19,] 119.48640 119.67038
#> [20,] 115.95258 115.79724
#> [21,]  94.92629  93.95212
#> [22,]  82.49848  79.38514
#> [23,]  70.20684  73.87151
#> [24,]  71.49825  78.00000
#> [25,]  69.28925  68.09552
#> [26,]  63.92183  67.80855
#> [27,]  44.29447  47.34976
#> [28,]  34.55431  41.88078
#> [29,]   0.00000  21.26029
#> [30,]  21.26029   0.00000