Calculate pairwise distances between two matrices or data frames.

This function takes two data frames or matrices and returns a matrix of pairwise Euclidean distances between the two.

Usage

pairDist(x1, x2 = NULL, na.rm = FALSE)

Arguments

x1

Data frame or matrix one or more columns wide.

x2

Data frame or matrix one or more columns wide. If NULL, then pairwise distances between all points in x1 are calculated.

na.rm

Logical, if TRUE then any rows in x1 or x2 with at least one NA are removed first.

Value

Matrix with nrow(x1) rows and nrow(x2) columns. Values are the distance between each row of x1 and row of x2.

See also

Examples

x1 <- data.frame(x=sample(1:30, 30), y=sort(round(100 * rnorm(30))), z=sample(1:30, 30))
x2 <- data.frame(x=1:20, y=round(100 * rnorm(20)), z=sample(1:20, 20))
pairDist(x1, x2)
#>            [,1]       [,2]      [,3]      [,4]     [,5]       [,6]      [,7]
#>  [1,] 212.84736 354.104504 144.27751 125.35549 573.4649 207.089353 279.14512
#>  [2,] 122.51531 262.769861  55.44367  40.57093 481.6368 116.112015 187.64328
#>  [3,]  68.60758 204.570281  25.67100  43.50862 423.2033  59.874870 129.78829
#>  [4,]  40.79216 182.466435  30.33150  49.61854 401.0511  37.121422 107.39646
#>  [5,]  44.66542 169.755707  52.41183  74.91996 387.8208  33.120990  95.85927
#>  [6,]  16.15549 144.086779  66.27971  89.76079 363.6193   6.403124  69.53416
#>  [7,]  28.44293 133.195345  79.98125 103.58089 352.2087  20.615528  59.15235
#>  [8,]  16.40122 131.768737  80.03124 100.72239 350.1614  20.124612  56.75385
#>  [9,]  30.56141 130.499042  83.29466 106.57392 349.2506  23.958297  56.60389
#> [10,]  26.94439 124.987999  89.45390 109.25200 342.3419  31.112698  50.93133
#> [11,]  32.95451 113.815640  97.78548 119.22248 332.3673  35.524639  38.88444
#> [12,]  49.61854 107.368524 109.13295 131.27833 324.5643  48.559242  36.68787
#> [13,]  52.09607  90.818500 120.50726 141.08862 309.0275  58.497863  18.27567
#> [14,]  56.32051  87.669835 125.04399 145.03103 305.0279  63.796552  19.54482
#> [15,]  71.40728  71.700767 139.26593 160.30284 290.1258  76.557168   7.28011
#> [16,]  80.56674  64.163853 147.41438 168.96449 282.4907  84.433406  13.89244
#> [17,]  88.85944  59.143892 155.04838 176.55028 275.8732  92.249661  23.62202
#> [18,]  98.80283  47.853944 165.69249 187.05080 264.6394 102.781321  31.96873
#> [19,] 101.13852  51.487863 166.74831 188.27905 265.4600 104.139330  36.18011
#> [20,] 102.78619  47.675990 169.35466 190.43372 261.9217 106.700515  37.24245
#> [21,] 116.94871  26.476405 184.06792 205.98301 245.7377 121.016528  49.26459
#> [22,] 125.57070  30.364453 191.31126 213.44320 240.9772 128.382242  58.32667
#> [23,] 133.40915  18.920888 201.57877 222.13059 228.0877 138.902844  67.20119
#> [24,] 149.06710   9.949874 217.08293 238.21209 212.2004 154.236831  82.24962
#> [25,] 155.13220  20.832667 223.51286 244.00410 206.0000 160.903698  89.12912
#> [26,] 175.41665  35.000000 243.21801 264.35015 186.4001 180.235956 108.18965
#> [27,] 177.23149  37.269290 245.26516 266.15973 184.1657 182.375985 110.33132
#> [28,] 203.71794  64.257295 270.89666 292.25160 161.0497 207.872557 136.20940
#> [29,] 213.63754  74.101282 281.69665 302.28629 148.4891 218.862971 146.91834
#> [30,] 226.24102  86.423376 293.19789 314.69827 140.5169 230.171675 158.64741
#>            [,8]      [,9]     [,10]      [,11]      [,12]     [,13]     [,14]
#>  [1,] 313.23154 201.72754 301.18101 286.872794 227.321358 256.56968 246.91092
#>  [2,] 221.55361 110.42645 209.40392 195.053839 135.502767 164.61470 154.93547
#>  [3,] 163.51453  55.79426 151.25145 137.764291  78.204859 107.15876  98.03061
#>  [4,] 141.13114  29.01724 129.25169 114.087686  55.524769  84.11896  74.19569
#>  [5,] 129.03100  33.48134 116.70904 104.359954  46.690470  74.10128  66.04544
#>  [6,] 103.65809  18.84144  91.70060  79.101201  21.679483  49.32545  42.00000
#>  [7,]  92.74158  30.51229  80.45496  69.260378  16.309506  39.56008  34.00000
#>  [8,]  90.22749  22.38303  78.28793  63.316664   6.480741  33.00000  23.36664
#>  [9,]  89.94443  32.80244  77.62731  66.407831  15.779734  36.85105  31.54362
#> [10,]  83.25863  32.09361  71.41428  55.767374  12.688578  26.26785  16.15549
#> [11,]  72.34639  40.89010  60.21628  46.497312  14.317821  15.93738  10.34408
#> [12,]  67.00746  54.93633  54.82700  45.011110  29.748950  22.04541  23.93742
#> [13,]  49.65884  63.22974  38.54867  23.280893  38.691084  12.40967  20.61553
#> [14,]  47.09565  67.83804  36.95944  20.688161  44.395946  18.76166  25.86503
#> [15,]  30.14963  82.03048  18.78829   8.185353  56.284989  27.47726  37.81534
#> [16,]  23.00000  90.49309  10.81665  14.247807  64.062470  35.70714  46.32494
#> [17,]  20.24846  98.15294  11.57584  20.976177  71.630999  43.49713  53.82379
#> [18,]  10.95445 108.60018  11.31371  26.870058  82.225300  53.57238  63.91400
#> [19,]  20.32240 110.04090  19.20937  32.264532  83.510478  55.72253  65.83312
#> [20,]  16.03122 112.12047  17.91647  30.740852  85.848704  57.21014  67.23095
#> [21,]  17.11724 127.69886  27.94638  45.967380 101.419919  73.32803  83.92854
#> [22,]  29.35984 135.07776  37.44329  54.990908 108.346666  80.67218  91.15372
#> [23,]  33.37664 144.06943  45.36518  59.008474 118.427193  89.16277  99.05049
#> [24,]  48.30114 160.14056  60.53924  75.848533 134.335401 105.49408 115.68059
#> [25,]  55.19058 166.15655  67.45369  81.228074 140.687597 111.50785 121.37133
#> [26,]  74.11478 186.06719  86.07555 101.405128 160.012500 131.03816 141.18782
#> [27,]  76.19055 188.00266  88.27797 103.126136 162.123410 133.03759 143.08739
#> [28,] 102.54267 213.83405 114.03070 129.634100 187.494000 158.71043 168.86089
#> [29,] 112.80514 224.19857 124.77580 139.129436 198.368344 169.15378 179.03631
#> [30,] 125.10396 236.26468 136.51007 152.266214 209.842798 181.17671 191.34785
#>           [,15]      [,16]     [,17]     [,18]     [,19]      [,20]
#>  [1,] 462.54621 175.789647 139.44892 253.57247 280.90034 293.671585
#>  [2,] 370.41193  83.701852  47.26521 161.24826 188.54442 201.258540
#>  [3,] 311.97115  30.016662  16.15549 103.26665 130.72873 143.024473
#>  [4,] 290.07758   8.602325  36.35932  81.65170 108.46658 121.536003
#>  [5,] 276.63333  26.476405  50.72475  69.45502  96.67471 108.332821
#>  [6,] 253.04940  39.560081  72.62920  47.18050  74.01351  85.918566
#>  [7,] 241.30064  50.328918  83.49251  35.41186  62.46599  73.607065
#>  [8,] 239.06066  48.135226  84.81156  30.56141  57.39338  70.356236
#>  [9,] 238.24567  52.896125  86.39444  32.26453  59.23681  70.349129
#> [10,] 231.05627  56.462377  93.26307  23.17326  49.02040  62.201286
#> [11,] 221.22613  66.075714 102.19589  12.52996  40.01250  52.392748
#> [12,] 213.33776  77.058419 111.61541  15.90597  36.12478  45.923850
#> [13,] 198.31541  89.944427 126.45157  18.78829  21.93171  33.361655
#> [14,] 194.52763  94.731199 131.36971  25.19921  22.20360  32.664966
#> [15,] 179.23727 108.300508 144.62711  31.68596  12.08305  15.620499
#> [16,] 171.42345 116.124933 152.08221  38.41875  15.68439   8.717798
#> [17,] 164.62381 123.247718 159.07860  45.37621  21.18962   8.660254
#> [18,] 153.42425 134.074606 170.07645  56.08030  30.38092  16.881943
#> [19,] 154.20117 134.688530 170.29680  57.14893  32.77194  19.646883
#> [20,] 150.53903 137.247951 173.28301  59.31273  33.13608  19.646883
#> [21,] 135.37356 153.652205 189.32248  76.48529  51.94228  39.268308
#> [22,] 130.29582 159.956244 195.21783  82.52878  58.07753  44.485953
#> [23,] 117.06836 170.217508 206.70752  92.61209  65.52099  53.113087
#> [24,] 101.95097 186.453748 222.62300 109.04128  82.71638  70.099929
#> [25,]  95.54580 192.525323 229.02402 115.17378  88.20998  75.927597
#> [26,]  75.48510 212.040091 248.17937 134.13426 107.39646  94.361009
#> [27,]  73.28028 214.105114 250.38570 136.29380 109.38921  96.540147
#> [28,]  50.99020 239.271812 275.12361 161.26376 134.66997 121.367211
#> [29,]  37.06751 250.129966 286.50480 172.23821 145.04482 132.219514
#> [30,]  34.58323 261.554966 297.24905 183.60283 157.13688 143.784561
pairDist(x1)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]       [,7]
#>  [1,]   0.00000  93.34881 152.19396 172.95086 188.47546 210.06904 222.218361
#>  [2,]  93.34881   0.00000  59.43904  82.08532  95.85927 119.72051 130.449224
#>  [3,] 152.19396  59.43904   0.00000  33.03029  36.44173  63.76519  71.763500
#>  [4,] 172.95086  82.08532  33.03029   0.00000  34.56877  41.82105  55.072679
#>  [5,] 188.47546  95.85927  36.44173  34.56877   0.00000  36.90528  37.709415
#>  [6,] 210.06904 119.72051  63.76519  41.82105  36.90528   0.00000  19.390719
#>  [7,] 222.21836 130.44922  71.76350  55.07268  37.70942  19.39072   0.000000
#>  [8,] 223.65375 131.76874  75.11325  51.19570  45.01111  22.04541  20.248457
#>  [9,] 225.33974 133.27415  74.35725  57.87918  39.61060  23.21637   4.358899
#> [10,] 232.22834 139.82131  82.92768  60.04998  51.58488  33.00000  25.942244
#> [11,] 241.38351 149.28496  91.46037  69.61322  58.28379  34.79943  24.228083
#> [12,] 250.88244 158.16447  99.04040  82.06095  63.31666  47.71792  30.870698
#> [13,] 264.69983 173.50504 116.84605  92.14120  84.74078  56.64804  49.365980
#> [14,] 269.13937 178.26385 122.22520  96.51943  90.79097  62.04031  55.901699
#> [15,] 283.39901 191.77330 134.17898 111.04504 100.49876  74.39086  64.389440
#> [16,] 291.22328 199.20090 140.86873 119.44455 106.15084  82.18272  70.242437
#> [17,] 298.43592 206.02427 147.36010 126.96456 112.12047  90.22749  77.103826
#> [18,] 309.35740 217.09445 158.58436 137.52818 123.50304 100.60318  88.141931
#> [19,] 309.70470 217.05990 158.15815 138.69751 122.53163 102.18610  88.260977
#> [20,] 312.69314 220.12042 161.53637 140.92906 126.29331 104.82366  91.760558
#> [21,] 328.02591 236.54386 178.19652 156.65567 143.31085 118.08472 106.770783
#> [22,] 334.38900 242.10328 183.02732 163.82918 147.19715 125.82925 112.236358
#> [23,] 345.66168 253.74397 195.81879 173.03757 161.24826 136.62723 125.614490
#> [24,] 361.29351 269.82031 211.78763 189.13752 177.10449 151.47937 140.762211
#> [25,] 367.72544 276.11230 218.34148 195.10510 183.87224 158.42664 147.901995
#> [26,] 387.25056 295.27784 236.82905 215.07906 201.61349 177.62320 165.921668
#> [27,] 389.38285 297.48445 239.21538 217.00461 204.17639 179.80545 168.422089
#> [28,] 414.45024 322.02019 263.12355 242.77768 227.37194 205.35335 192.525323
#> [29,] 425.68650 333.47114 275.11089 253.15805 239.87914 216.44861 204.660695
#> [30,] 436.56500 344.07703 285.04386 265.18107 249.13049 227.60931 214.508741
#>            [,8]       [,9]     [,10]     [,11]     [,12]      [,13]      [,14]
#>  [1,] 223.65375 225.339743 232.22834 241.38351 250.88244 264.699830 269.139369
#>  [2,] 131.76874 133.274154 139.82131 149.28496 158.16447 173.505043 178.263849
#>  [3,]  75.11325  74.357246  82.92768  91.46037  99.04040 116.846053 122.225202
#>  [4,]  51.19570  57.879185  60.04998  69.61322  82.06095  92.141196  96.519428
#>  [5,]  45.01111  39.610605  51.58488  58.28379  63.31666  84.740781  90.790969
#>  [6,]  22.04541  23.216374  33.00000  34.79943  47.71792  56.648036  62.040309
#>  [7,]  20.24846   4.358899  25.94224  24.22808  30.87070  49.365980  55.901699
#>  [8,]   0.00000  20.024984  11.35782  18.78829  34.33657  42.083251  47.254629
#>  [9,]  20.02498   0.000000  23.79075  21.63331  26.87006  47.222876  53.870214
#> [10,]  11.35782  23.790755   0.00000  14.35270  28.03569  36.687873  41.785165
#> [11,]  18.78829  21.633308  14.35270   0.00000  18.70829  26.720778  33.075671
#> [12,]  34.33657  26.870058  28.03569  18.70829   0.00000  33.105891  39.774364
#> [13,]  42.08325  47.222876  36.68787  26.72078  33.10589   0.000000   7.071068
#> [14,]  47.25463  53.870214  41.78516  33.07567  39.77436   7.071068   0.000000
#> [15,]  60.21628  61.717096  53.60037  42.76681  41.04875  19.824228  18.841444
#> [16,]  68.36666  67.268120  61.29437  50.04998  44.05678  30.479501  30.380915
#> [17,]  75.84853  73.851202  68.07349  57.39338  49.01020  39.572718  39.446166
#> [18,]  86.35392  85.005882  78.66384  68.07349  60.51446  47.979162  46.540305
#> [19,]  87.70405  84.858706  79.55501  69.23150  59.21993  52.182373  51.778374
#> [20,]  89.76636  88.425110  81.49233  71.57514  63.09517  52.469038  51.078371
#> [21,] 105.91506 104.052871  99.30257  87.74395  81.09871  65.764732  63.442888
#> [22,] 112.89376 109.077954 105.36603  94.24436  84.20214  75.511589  74.269779
#> [23,] 122.11879 122.637678 114.35908 104.55620  98.55963  81.326502  77.704569
#> [24,] 138.37630 137.989130 131.24405 120.63582 114.52074  97.133928  93.557469
#> [25,] 144.39183 145.037926 136.85759 126.96456 121.25180 103.082491  99.085821
#> [26,] 164.03048 162.938639 156.36176 146.04451 138.45216 123.405835 120.012499
#> [27,] 166.03012 165.459965 158.34456 148.20594 141.07090 125.159898 121.536003
#> [28,] 191.58810 189.359447 183.52929 173.33494 164.13714 151.937487 148.932871
#> [29,] 202.12867 201.566366 194.03350 184.33936 176.59841 161.551849 157.825853
#> [30,] 213.99533 211.322029 205.90046 195.66553 186.01344 174.542258 171.607109
#>           [,15]      [,16]      [,17]      [,18]      [,19]      [,20]
#>  [1,] 283.39901 291.223282 298.435923 309.357398 309.704698 312.693140
#>  [2,] 191.77330 199.200904 206.024270 217.094449 217.059900 220.120422
#>  [3,] 134.17898 140.868733 147.360103 158.584362 158.158149 161.536374
#>  [4,] 111.04504 119.444548 126.964562 137.528179 138.697513 140.929060
#>  [5,] 100.49876 106.150836 112.120471 123.503036 122.531629 126.293309
#>  [6,]  74.39086  82.182723  90.227490 100.603181 102.186105 104.823661
#>  [7,]  64.38944  70.242437  77.103826  88.141931  88.260977  91.760558
#>  [8,]  60.21628  68.366659  75.848533  86.353923  87.704048  89.766363
#>  [9,]  61.71710  67.268120  73.851202  85.005882  84.858706  88.425110
#> [10,]  53.60037  61.294372  68.073490  78.663842  79.555012  81.492331
#> [11,]  42.76681  50.049975  57.393379  68.073490  69.231496  71.575135
#> [12,]  41.04875  44.056782  49.010203  60.514461  59.219929  63.095166
#> [13,]  19.82423  30.479501  39.572718  47.979162  52.182373  52.469038
#> [14,]  18.84144  30.380915  39.446166  46.540305  51.778374  51.078371
#> [15,]   0.00000  11.661904  21.047565  28.301943  33.436507  33.256578
#> [16,]  11.66190   0.000000   9.949874  18.466185  22.405357  23.452079
#> [17,]  21.04757   9.949874   0.000000  11.575837  12.688578  14.866069
#> [18,]  28.30194  18.466185  11.575837   0.000000   9.848858   7.000000
#> [19,]  33.43651  22.405357  12.688578   9.848858   0.000000   7.348469
#> [20,]  33.25658  23.452079  14.866069   7.000000   7.348469   0.000000
#> [21,]  46.36809  38.183766  33.778692  23.259407  28.390139  25.377155
#> [22,]  55.81219  45.287967  37.416574  28.178006  26.551836  26.095977
#> [23,]  62.36185  55.398556  49.779514  38.288379  42.154478  36.592349
#> [24,]  78.16649  71.035203  65.795137  54.267854  57.827329  53.347915
#> [25,]  84.43341  77.801028  72.401657  60.844063  64.319515  59.219929
#> [26,] 104.09611  96.104110  89.593527  78.287930  79.737068  75.960516
#> [27,] 106.03773  98.407317  92.103203  80.703160  82.522724  78.421936
#> [28,] 132.34047 123.507085 116.073253 105.275828 105.023807 102.127371
#> [29,] 142.45701 134.662541 127.831921 116.623325 117.524466 113.657380
#> [30,] 154.89997 145.897224 138.322088 127.667537 127.051171 124.426685
#>           [,21]     [,22]     [,23]     [,24]     [,25]      [,26]      [,27]
#>  [1,] 328.02591 334.38900 345.66168 361.29351 367.72544 387.250565 389.382845
#>  [2,] 236.54386 242.10328 253.74397 269.82031 276.11230 295.277835 297.484453
#>  [3,] 178.19652 183.02732 195.81879 211.78763 218.34148 236.829052 239.215384
#>  [4,] 156.65567 163.82918 173.03757 189.13752 195.10510 215.079055 217.004608
#>  [5,] 143.31085 147.19715 161.24826 177.10449 183.87224 201.613492 204.176394
#>  [6,] 118.08472 125.82925 136.62723 151.47937 158.42664 177.623197 179.805450
#>  [7,] 106.77078 112.23636 125.61449 140.76221 147.90199 165.921668 168.422089
#>  [8,] 105.91506 112.89376 122.11879 138.37630 144.39183 164.030485 166.030118
#>  [9,] 104.05287 109.07795 122.63768 137.98913 145.03793 162.938639 165.459965
#> [10,]  99.30257 105.36603 114.35908 131.24405 136.85759 156.361760 158.344561
#> [11,]  87.74395  94.24436 104.55620 120.63582 126.96456 146.044514 148.205938
#> [12,]  81.09871  84.20214  98.55963 114.52074 121.25180 138.452158 141.070904
#> [13,]  65.76473  75.51159  81.32650  97.13393 103.08249 123.405835 125.159898
#> [14,]  63.44289  74.26978  77.70457  93.55747  99.08582 120.012499 121.536003
#> [15,]  46.36809  55.81219  62.36185  78.16649  84.43341 104.096109 106.037729
#> [16,]  38.18377  45.28797  55.39856  71.03520  77.80103  96.104110  98.407317
#> [17,]  33.77869  37.41657  49.77951  65.79514  72.40166  89.593527  92.103203
#> [18,]  23.25941  28.17801  38.28838  54.26785  60.84406  78.287930  80.703160
#> [19,]  28.39014  26.55184  42.15448  57.82733  64.31951  79.737068  82.522724
#> [20,]  25.37716  26.09598  36.59235  53.34791  59.21993  75.960516  78.421936
#> [21,]   0.00000  19.00000  24.55606  34.61214  43.39355  59.983331  62.498000
#> [22,]  19.00000   0.00000  28.17801  37.21559  45.14421  56.169387  59.623821
#> [23,]  24.55606  28.17801   0.00000  19.20937  22.89105  42.766810  44.192760
#> [24,]  34.61214  37.21559  19.20937   0.00000  11.00000  27.531800  29.017236
#> [25,]  43.39355  45.14421  22.89105  11.00000   0.00000  23.430749  23.345235
#> [26,]  59.98333  56.16939  42.76681  27.53180  23.43075   0.000000   4.898979
#> [27,]  62.49800  59.62382  44.19276  29.01724  23.34524   4.898979   0.000000
#> [28,]  88.12491  80.52950  71.78440  57.81003  53.52569  30.561414  30.822070
#> [29,]  99.26732  94.30270  80.23092  66.00000  59.23681  39.370039  37.013511
#> [30,] 110.42645 102.29858  94.49339  80.13738  75.74299  52.820451  52.592775
#>           [,28]     [,29]     [,30]
#>  [1,] 414.45024 425.68650 436.56500
#>  [2,] 322.02019 333.47114 344.07703
#>  [3,] 263.12355 275.11089 285.04386
#>  [4,] 242.77768 253.15805 265.18107
#>  [5,] 227.37194 239.87914 249.13049
#>  [6,] 205.35335 216.44861 227.60931
#>  [7,] 192.52532 204.66069 214.50874
#>  [8,] 191.58810 202.12867 213.99533
#>  [9,] 189.35945 201.56637 211.32203
#> [10,] 183.52929 194.03350 205.90046
#> [11,] 173.33494 184.33936 195.66553
#> [12,] 164.13714 176.59841 186.01344
#> [13,] 151.93749 161.55185 174.54226
#> [14,] 148.93287 157.82585 171.60711
#> [15,] 132.34047 142.45701 154.89997
#> [16,] 123.50708 134.66254 145.89722
#> [17,] 116.07325 127.83192 138.32209
#> [18,] 105.27583 116.62333 127.66754
#> [19,] 105.02381 117.52447 127.05117
#> [20,] 102.12737 113.65738 124.42669
#> [21,]  88.12491  99.26732 110.42645
#> [22,]  80.52950  94.30270 102.29858
#> [23,]  71.78440  80.23092  94.49339
#> [24,]  57.81003  66.00000  80.13738
#> [25,]  53.52569  59.23681  75.74299
#> [26,]  30.56141  39.37004  52.82045
#> [27,]  30.82207  37.01351  52.59278
#> [28,]   0.00000  20.24846  22.71563
#> [29,]  20.24846   0.00000  25.01999
#> [30,]  22.71563  25.01999   0.00000